Added new Requests

+++ TableTamer/Models/Table.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public bool Status { get; set; } 
-        public float Position { get; set; }
+        public int Position { get; set; }
         
         public ICollection<Fatura>? Faturat { get; set; }
         public User? user { get; set; }
